requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1

# 데이터 (pandas-ta와 호환되는 pandas 버전은 pip가 해석)
yfinance==0.2.40
pandas>=2.2.2
pandas-ta==0.4.67b0

# 차트
matplotlib==3.9.0

# 뉴스
feedparser==6.0.11

# AI (httpx 최신과 호환)
anthropic>=0.40.0
